Return error when package not exist

+++ compiler-cli/src/remove.rs
@@ -19,6 +19,7 @@ pub fn command(packages: Vec<String>) -> Result<()> {
         })?;
 
     // Remove the specified dependencies
+    let mut packages_not_exist = vec![];
     for package_to_remove in packages.iter() {
         #[allow(clippy::indexing_slicing)]
         let maybe_removed_item = toml["dependencies"]
@@ -30,17 +31,22 @@ pub fn command(packages: Vec<String>) -> Result<()> {
             .as_table_like_mut()
             .and_then(|deps| deps.remove(package_to_remove));
 
-        if maybe_removed_item.or(maybe_removed_dev_item).is_some() {
-            cli::print_removed(package_to_remove);
-        } else {
-            cli::print_not_exist(package_to_remove);
+        if maybe_removed_item.or(maybe_removed_dev_item).is_none() {
+            packages_not_exist.push(package_to_remove.into());
         }
+    }
+
+    if !packages_not_exist.is_empty() {
+        return Err(Error::RemovedPackagesNotExist { packages: packages_not_exist });
     }
 
     // Write the updated config
     fs::write(Utf8Path::new("gleam.toml"), &toml.to_string())?;
     let paths = crate::find_project_paths()?;
     _ = crate::dependencies::download(&paths, cli::Reporter::new(), None, UseManifest::Yes)?;
+    for package_to_remove in packages {
+        cli::print_removed(&package_to_remove);
+    }
 
     Ok(())
 }
